NXT-511 Update paths to modules in unit tests for renamed pages
