+++ web/src/pages/__tests__/search.test.tsx
@@ -12,7 +12,7 @@ import sampleDataFailed from "@/mockData/search/search-results-failed.json";
 import sampleData from "@/mockData/search/search-results.json";
 import { render, screen, waitFor } from "@/test-utils/rendering";
 
-import SearchPage from "../search";
+import SearchPage from "../search.page";
 
 (useRouter as jest.Mock).mockImplementation(() => ({
 	route: "/",
